Make a ZarrV2 class. Move ZarrV2 and ZarrV3 into their own directories.
