+++ src/zarr.blosc.cpp
@@ -1,5 +1,6 @@
 #include "zarr.blosc.hh"
-#include "zarr.hh"
+#include "v2/zarr.v2.hh"
+#include "v3/zarr.v3.hh"
 
 #include "logger.h"
 
@@ -12,12 +13,12 @@ using json = nlohmann::json;
 namespace {
 template<zarr::BloscCodecId CodecId>
 struct Storage*
-compressed_zarr_init()
+compressed_zarr_v2_init()
 {
     try {
         zarr::CompressionParams params(
           zarr::compression_codec_as_string<CodecId>(), 1, 1);
-        return new zarr::Zarr(std::move(params));
+        return new zarr::ZarrV2(std::move(params));
     } catch (const std::exception& exc) {
         LOGE("Exception: %s\n", exc.what());
     } catch (...) {
@@ -105,16 +106,14 @@ zarr::BloscEncoder::flush_impl()
 }
 
 extern "C" {
-    struct Storage*
-    compressed_zarr_zstd_init()
+    struct Storage* compressed_zarr_v2_zstd_init()
     {
-        return compressed_zarr_init<zarr::BloscCodecId::Zstd>();
+        return compressed_zarr_v2_init<zarr::BloscCodecId::Zstd>();
     }
 
-    struct Storage*
-    compressed_zarr_lz4_init()
+    struct Storage* compressed_zarr_v2_lz4_init()
     {
-        return compressed_zarr_init<zarr::BloscCodecId::Lz4>();
+        return compressed_zarr_v2_init<zarr::BloscCodecId::Lz4>();
     }
 
     struct Storage*
